Fix broken unittests in Python 2.7.
Originally all test classes were decorated by going over the class via
inspection - however, this doesn't work in Python 2.7 because the
methods that are returned from getmembers are all unbound (in Python 3+,
this is fixed as all methods are just function objects).

Now this uses a metaclass with six to work in both Python 2 and 3.