Updated to install the latest version of requests when library is installed (#57)
Ensure that when the library is installed, the current latest version of requests (2.14.2) is installed.
If this is not specified in setup.py, then the library would only keep the previous version and would not upgrade. Also updated the version tag to point to a revision where the requirements.txt had requests specified as the appropriate version.